[#148] Update modules for services
