+++ app/services/deployment/helpers/push_code_to_server.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 
-require "github_app_client"
-
 module Deployment
   module Helpers
     class PushCodeToServer
