+++ app/services/deployment/helpers/repository_cloner.rb
@@ -4,8 +4,8 @@ module Deployment
   module Helpers
     class RepositoryCloner
       PROVIDER_REPOSITORY_URI = {
-        ProjectsConstants::Providers::GITHUB => ->(configuration) { GithubAppClient.new(configuration.project_integration_id).repo_uri(configuration.repo_path) },
-        ProjectsConstants::Providers::GITLAB => ->(configuration) { GitlabIntegration::ClientWrapper.new.clone_repository_uri(configuration.repo_path) }
+        ProjectsConstants::Providers::GITHUB => ->(configuration) { ::ProviderAPI::Github::AppClient.new(configuration.project_integration_id).repo_uri(configuration.repo_path) },
+        ProjectsConstants::Providers::GITLAB => ->(configuration) { ::ProviderAPI::Gitlab::BotClient.new.clone_repository_uri(configuration.repo_path) }
       }.freeze
 
       def initialize(repo_configuration)
