+++ app/services/gitlab_integration/merge_request_notificator.rb
@@ -8,7 +8,7 @@ module GitlabIntegration
 
     def call(text)
       repository = @project_instance.project.repositories.find_by(path: @project_instance.attached_repo_path)
-      merge_request = GitlabIntegration::MergeRequestApi.new(repository.integration_id, @project_instance.attached_pull_request_number)
+      merge_request = GitlabIntegration::MergeRequest.new(repository.integration_id, @project_instance.attached_pull_request_number)
       merge_request.update_info_comment(text)
     end
   end
