--- lib/github_user_client.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-class GithubUserClient
-  delegate :find_user_installations, to: :@client
-
-  def initialize(user_token)
-    @client = Octokit::Client.new(access_token: user_token)
-  end
-end
